backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incharj-backend-py"
version = "0.1.0"
description = "Parallel FastAPI migration target for Incharj backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "sqlalchemy>=2.0.40",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "alembic>=1.16.0",
    "celery>=5.5.0",
    "redis>=6.2.0",
    "httpx>=0.28.0",
    "tenacity>=9.1.0",
    "structlog>=25.4.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.0",
    "orjson>=3.11.0",
    "email-validator>=2.2.0",
    "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.0",
    "ruff>=0.12.0",
    "mypy>=1.17.0",
    "black>=25.1.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
